refactor packages, added javadoc
