+++ src/test/java/ru/postgrespro/perf/pgmicrobench/statanalyzer/tests/modality/dataSets/ModalityTestData.java
@@ -1,7 +1,6 @@
-package ru.postgrespro.perf.pgmicrobench.statanalyzer.Tests.TestDataSets;
+package ru.postgrespro.perf.pgmicrobench.statanalyzer.tests.modality.dataSets;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 public class ModalityTestData {
@@ -11,9 +10,7 @@ public class ModalityTestData {
 
     public ModalityTestData(String name, double[] values, int expectedModality) {
         this.name = name;
-        this.values = Collections.unmodifiableList(Arrays.asList(
-                Arrays.stream(values).boxed().toArray(Double[]::new)
-        ));
+        this.values = List.of(Arrays.stream(values).boxed().toArray(Double[]::new));
         this.expectedModality = expectedModality;
     }
 
@@ -31,7 +28,7 @@ public class ModalityTestData {
 
     @Override
     public String toString() {
-        return String.format("f.Tests.Modality.TestDataSets.ModalityTestData{name='%s', expectedModality=%d, values=%s}",
+        return String.format("f.test.modality.dataSets.ModalityTestData{name='%s', expectedModality=%d, values=%s}",
                 name, expectedModality, values);
     }
 }
